nots_nids/requirements.txt
ripser==0.6.8
persim==0.3.7
pot==0.9.3
umap-learn==0.5.6
cvxpy==1.4.2
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.1
joblib==1.3.2
torch==2.1.2
tqdm==4.66.1
memory-profiler==0.61.0
